Calling the method f_updateAllFlows of an energy asset now also updates the v_currentPowerX variables in the parent gridconnection!
This is a more fool-proof solution, as manually adding lines of code to add power flows of each energy asset is no longer needed. Now the f_updateAllFlows method is what triggers the timestep in the energyAsset itself and also immediately applies its power to the gridconnection.
(Only exception is energy assets connected to GridNodes; like the thermal mass of a district heating network.