Added method curtailElectricityConsumption to J_EAProfile and J_EAConsumption.
When calling the consumption curtailment function, specific accounting 'posts' like v_evChargingPowerElectric_kW are not updated accordingly. That would require a 'mirror' function of f_operateFixedAssets to also correct all the subtotals.
J_EAProfile and J_EAConsumption keep track of loadLoad_kWh; the consumption that was 'cut off'. This could be used later either for KPI calculation or to implement a 'catch-up-consumption' method.