Holon2 developments gillis (#38)
* Added f_batteryManagementNodalPricing to GridConnection, using GridNode nodal pricing

* Made J_EAProfile compatible with profiles of arbitrary time resolution. Constructor now takes profileTimestep_h as input, instead of model timestep_h

* Forgot one replacement of timestep_h with profileTimestep_h in J_EAProfile

* Added method curtailElectricityConsumption to J_EAProfile and J_EAConsumption.

When calling the consumption curtailment function, specific accounting 'posts' like v_evChargingPowerElectric_kW are not updated accordingly. That would require a 'mirror' function of f_operateFixedAssets to also correct all the subtotals.
J_EAProfile and J_EAConsumption keep track of loadLoad_kWh; the consumption that was 'cut off'. This could be used later either for KPI calculation or to implement a 'catch-up-consumption' method.

* Removed time(HOUR) and time(MINUTE) statements where needed.

* Superbasic district heating reinstated, without use of transportBuffer (storageHeat) asset in the GridNodeHeat

* Commit before merge, disabled DLR for other projects.