bugfixes for when a coop is made using the filter functionality
Added a functionality to the accumulator.add method. Where it is now also possible to add an accumulator with a smaller resolution to an accumulator with a larger resolution if it matches time intervals.
The data resolution of the rapid run class of the  coop after creation is now equal to the largest interval of the members. When a coop is present during rapid run sim, the resolution of the coop is converted into more detail again.